+++ src/views/blocks/ImageBlock.php
@@ -20,23 +20,26 @@ $align = $this->varValue('align', 'left');
 <?php if ($image = $this->extraValue('image', false)): ?>
     <div class="image">
         <figure class="figure d-block <?php if ($align === 'left'): ?>text-left<?php elseif ($align === 'right'): ?>text-right<?php else: ?>text-center<?php endif; ?>">
-            <?php if ($this->cfgValue('lazyload', false)):
+            <?php
+            $caption = !empty($image->caption) ? $image->caption : '';
+            if ($this->cfgValue('lazyload', false)):
+                $options['src'] = 'data:image/gif;base64,R0lGODlhAQABAIAAAP///wAAACH5BAEAAAAALAAAAAABAAEAAAICRAEAOw==';
+                if (!empty($caption)) {
+                    $options['alt'] = $caption;
+                    $options['title'] = $caption;
+                }
                 echo LazyLoad::widget([
                     'src' => $image->source,
                     'extraClass' => 'figure-img img-fluid',
                     'width' => $image->itemArray['resolution_width'],
                     'height' => $image->itemArray['resolution_height'],
-                    'options' => [
-                        'src' => 'data:image/gif;base64,R0lGODlhAQABAIAAAP///wAAACH5BAEAAAAALAAAAAABAAEAAAICRAEAOw==',
-                        'alt' => $image->caption,
-                        'title' => $image->caption,
-                    ]
+                    'options' => $options
                 ]);
             else: ?>
-                <img src="<?= $this->extraValue('image')->source ?>" class="figure-img img-fluid" <?php if ($this->extraValue('image')->caption): ?>alt="<?= $this->extraValue('image')->caption ?>"<?php endif; ?>>
+                <img src="<?= $image->source ?>" class="figure-img img-fluid"<?php if (!empty($caption)): ?> alt="<?= $caption ?>"<?php endif; ?>>
             <?php endif; ?>
-            <?php if ($this->varValue('showCaption', false) && $this->extraValue('image')->caption): ?>
-                <figcaption class="figure-caption"><?= $this->extraValue('image')->caption ?></figcaption>
+            <?php if ($this->varValue('showCaption', false) && !empty($caption)): ?>
+                <figcaption class="figure-caption"><?= $caption ?></figcaption>
             <?php endif; ?>
         </figure>
     </div>